COMP: getting rid of uninit var, casting compiler warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12696
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12696 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/UtahBSpline/UtahBSplineDeformableRegistration.cxx
@@ -374,7 +374,7 @@ template<class T> int DoIt( int argc, char * argv[], T )
   int ncps[3]; 
   int degree[3]; 
   for(int i=0; i < 3; i++){
-    ncps[i] = floor(((float)imageSize[i])/gridSize);
+    ncps[i] = (int)floor(((float)imageSize[i])/gridSize);
     degree[i] = 2;
 
     if(ncps[i] < degree[i] + 2){
